split some classes and stucts off looking for character levels
