+++ FF8/FF8.csproj
@@ -161,7 +161,13 @@
     <Compile Include="Menu\Module_main_menu_Loading.cs" />
     <Compile Include="Menu\Module_main_menu_MainLobby.cs" />
     <Compile Include="Menu\Module_main_menu_DebugLobby.cs" />
+    <Compile Include="Saves.CharacterData.cs" />
+    <Compile Include="Saves.Characters.cs" />
     <Compile Include="Saves.cs" />
+    <Compile Include="Saves.Data.cs" />
+    <Compile Include="Saves.FF8Variables.cs" />
+    <Compile Include="Saves.GFData.cs" />
+    <Compile Include="Saves.GFs.cs" />
     <Compile Include="Sources\Environment\Game\Location\GameLocation.cs" />
     <Compile Include="Sources\Environment\Game\Location\IGameLocationProvider.cs" />
     <Compile Include="Sources\Environment\Game\Location\LinuxGameLocationProvider.cs" />
