+++ FF8/Saves.Characters.cs
@@ -1,0 +1,18 @@
+﻿namespace FF8
+{
+
+
+
+
+
+
+    internal static partial class Saves
+    {
+        internal enum Characters
+        {
+            // I noticed some values were in order of these characters so I made those values into arrays
+            // and put the character names into an enum.
+            Squall, Zell, Irvine, Quistis, Rinoa, Selphie, Seifer, Edea, Count // Count is the last value and is the number of characters in enum
+        }
+    }
+}
