+++ FF8/Saves.GFs.cs
@@ -1,0 +1,32 @@
+﻿namespace FF8
+{
+
+
+
+
+
+
+    internal static partial class Saves
+    {
+        internal enum GFs
+        {
+            Quezacotl,
+            Shiva,
+            Ifrit,
+            Siren,
+            Brothers,
+            Diablos,
+            Carbuncle,
+            Leviathan,
+            Pandemona,
+            Cerberus,
+            Alexander,
+            Doomtrain,
+            Bahamut,
+            Cactuar,
+            Tonberry,
+            Eden,
+            Count
+        }
+    }
+}
